app-engine/GCP-App-Engine-flexible/ML-model/credit-card-default/data-validation/requirements.txt
Flask==2.1.0
google-cloud-bigquery
numpy 
scikit-learn
gunicorn==20.1.0
gcsfs==0.6.2
pandas-gbq